Start creating String types and segment DataTypes
Create String Class along with VarChar and Char. Separated DataTypes
into new files and improved file loading logic.
Implement beginning of hierarchal order of DataTypes. This hierarchy
will be used when validating incoming data in conjunction with is_a()
where we will check the expected data type and see if the incoming value
is a subclass of that type.